# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerOS API",
    "description": "Private career profiles, evidence-based eligibility and opportunity tracking. Cookie auth; mutations require X-CSRF-Token.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/profile": {
      "get": {
        "tags": [
          "Career profile"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Career profile"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/skills": {
      "get": {
        "tags": [
          "Career profile"
        ],
        "summary": "Skills",
        "operationId": "skills_skills_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Career profile"
        ],
        "summary": "Add Skill",
        "operationId": "add_skill_skills_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/skills/{id}": {
      "delete": {
        "tags": [
          "Career profile"
        ],
        "summary": "Remove Skill",
        "operationId": "remove_skill_skills__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects": {
      "get": {
        "tags": [
          "Career profile"
        ],
        "summary": "Projects",
        "operationId": "projects_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Career profile"
        ],
        "summary": "Add Project",
        "operationId": "add_project_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}": {
      "put": {
        "tags": [
          "Career profile"
        ],
        "summary": "Edit Project",
        "operationId": "edit_project_projects__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Career profile"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project_projects__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/mastery": {
      "put": {
        "tags": [
          "Career profile"
        ],
        "summary": "Edit Mastery",
        "operationId": "edit_mastery_projects__id__mastery_put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasteryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resumes": {
      "get": {
        "tags": [
          "Career profile"
        ],
        "summary": "Resumes",
        "operationId": "resumes_resumes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Career profile"
        ],
        "summary": "Add Resume",
        "operationId": "add_resume_resumes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resumes/{id}": {
      "put": {
        "tags": [
          "Career profile"
        ],
        "summary": "Edit Resume",
        "operationId": "edit_resume_resumes__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Career profile"
        ],
        "summary": "Delete Resume",
        "operationId": "delete_resume_resumes__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MasteryIn": {
        "properties": {
          "topic": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Topic"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "LEARNING",
              "COMFORTABLE",
              "STRONG"
            ],
            "title": "Status"
          },
          "confidence": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 1
          },
          "notes": {
            "type": "string",
            "maxLength": 5000,
            "title": "Notes",
            "default": ""
          },
          "linked_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Linked Skills"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "topic",
          "status"
        ],
        "title": "MasteryIn"
      },
...
      "PreferenceIn": {
        "properties": {
          "target_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Target Roles"
          },
          "countries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Countries"
          },
          "watchlist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Watchlist"
          },
          "preferred_companies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Preferred Companies"
          },
          "deadline_days": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Deadline Days"
          },
          "email_enabled": {
            "type": "boolean",
            "title": "Email Enabled",
            "default": false
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Weights"
          },
          "weekly_plan": {
            "type": "string",
            "maxLength": 5000,
            "title": "Weekly Plan",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PreferenceIn"
      },
      "ProfileIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "university": {
            "type": "string",
            "maxLength": 200,
            "title": "University",
            "default": ""
          },
          "degree": {
            "type": "string",
            "maxLength": 120,
            "title": "Degree",
            "default": ""
          },
          "branch": {
            "type": "string",
            "maxLength": 120,
            "title": "Branch",
            "default": ""
          },
          "graduation_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 1980.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduation Year"
          },
          "cgpa": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cgpa"
          },
          "semester": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "experience_years": {
            "type": "number",
            "maximum": 60.0,
            "minimum": 0.0,
            "title": "Experience Years",
            "default": 0
          },
          "work_authorizations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Work Authorizations"
          },
          "external_profiles": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "maxProperties": 10,
            "title": "External Profiles"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferenceIn"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProfileIn"
      },
      "ProjectIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "title": "Description",
            "default": ""
          },
          "repository_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Url"
          },
          "live_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Url"
          },
          "technologies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Technologies"
          },
          "verified_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Verified Skills"
          },
          "interview_readiness": {
            "type": "string",
            "maxLength": 80,
            "title": "Interview Readiness",
            "default": "NOT_STARTED"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectIn"
      },
...
      "ResumeIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "is_master": {
            "type": "boolean",
            "title": "Is Master",
            "default": false
          },
          "file_name": {
            "type": "string",
            "maxLength": 200,
            "title": "File Name",
            "default": ""
          },
          "notes": {
            "type": "string",
            "maxLength": 5000,
            "title": "Notes",
            "default": ""
          },
          "version": {
            "type": "string",
            "maxLength": 30,
            "title": "Version",
            "default": "1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ResumeIn"
      },
      "SkillIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "category": {
            "type": "string",
            "maxLength": 80,
            "title": "Category",
            "default": "Programming"
          },
          "proficiency": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Proficiency",
            "default": 1
          },
          "confidence": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 1
          },
          "evidence": {
            "type": "string",
            "maxLength": 3000,
            "title": "Evidence",
            "default": ""
          },
          "last_used": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used"
          },
          "learning_status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "LEARNING",
              "COMFORTABLE",
              "STRONG"
            ],
            "title": "Learning Status",
            "default": "LEARNING"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SkillIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}